% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logit_link.R
\encoding{UTF-8}
\name{logit_link}
\alias{logit_link}
\title{calculate the logistic link function}
\usage{
logit_link(x)
}
\arguments{
\item{x}{numeric vector. Values to calculate the logistic link function for}
}
\description{
calculate the logistic link function for one or more observations
}
\author{
Janek Thomas, Philipp Rösch
}

