% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logitreg.list.R
\name{logitreg.list}
\alias{logitreg.list}
\title{Calculate a logistic regression for list design input}
\usage{
\method{logitreg}{list}(design, response, ...)
}
\arguments{
\item{design}{list. Design matrix containing the observations, first
column should be constant 1 as intercept}

\item{response}{numeric vector. Response vector of 1s and 0s, should have the
same length as the number of rows in design.
#'}
}
\value{
A list with estimated coefficients, fitted propabilities and
original data
}
\description{
Calculate a logistic regression for given data. Uses
\code{\link[stats]{optim}} with the Broyden-Fletcher-Goldfarb-Shanno (BFGS)
algorithm as default for parameter estimation
}
\author{
Janek Thomas, Philipp Roesch
}
\seealso{
The default method \code{\link{logitreg.default}} and  \code{\link{logitreg.list}} for lists
}

