% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logitreg.R
\encoding{UTF-8}
\name{logitreg}
\alias{logitreg}
\title{Calculate a logistic regression}
\usage{
logitreg(design, ...)
}
\arguments{
\item{design}{numeric matrix or list containing the observations. First
column/list should be constant 1 as intercept. resonse need to be specifies.
Alternativly an object of class "formula".}

\item{...}{Further parameters passed to other logitreg inout and
\code{\link[stats]{optim}}.}
}
\value{
A list with estimated coefficients, fitted propabilities and
original data.
}
\description{
Generic function to calculate a logistic regression for given data. Uses
\code{\link[stats]{optim}} with the Broyden-Fletcher-Goldfarb-Shanno (BFGS)
algorithm as default for parameter estimation.
}
\author{
Janek Thomas, Philipp Rösch
}
\seealso{
The default method \code{\link{logitreg.default}} for matrixes
(and data frames), \code{\link{logitreg.formula}} for formula, and
\code{\link{logitreg.list}} for lists
}

